+++ globalshop-order-app/src/main/java/com/wangqin/globalshop/order/app/uitl/ErpOrderUtil.java
@@ -36,6 +36,6 @@ public class ErpOrderUtil {
 				}
 			}
 		}
-		erpOrder.setStockStatus((byte) stockstatus);
+		erpOrder.setStockStatus( stockstatus);
 	}
 }
